+++ Konsole/Case.cs
@@ -1,0 +1,4 @@
+﻿namespace Konsole
+{
+    public enum Case {  Sensitive, Insensitive }
+}